refactor(fieldmap): extract SVG label layer to its own component
